spirv: don't leak spirv headers downstream
