+++ src/thorin/be/spirv/spirv.h
@@ -1,14 +1,13 @@
 #ifndef THORIN_SPIRV_H
 #define THORIN_SPIRV_H
 
-#include "thorin/be/spirv/spirv_builder.hpp"
 #include "thorin/analyses/schedule.h"
 #include "thorin/be/codegen.h"
 #include "thorin/be/emitter.h"
 
 namespace thorin::spirv {
 
-using SpvId = builder::SpvId;
+using SpvId = uint32_t;
 
 class CodeGen;
 
@@ -35,25 +34,6 @@ struct ConvertedType {
     std::optional<Layout> layout;
 };
 
-struct BasicBlockBuilder : public builder::SpvBasicBlockBuilder {
-    explicit BasicBlockBuilder(FnBuilder& fn_builder);
-    BasicBlockBuilder(const BasicBlockBuilder&) = delete;
-
-    FnBuilder& fn_builder;
-    FileBuilder& file_builder;
-    std::unordered_map<const Param*, Phi> phis_map;
-    DefMap<SpvId> args;
-};
-
-struct FnBuilder : public builder::SpvFnBuilder {
-    explicit FnBuilder(FileBuilder& file_builder);
-    FnBuilder(const FnBuilder&) = delete;
-
-    FileBuilder& file_builder;
-    std::vector<std::unique_ptr<BasicBlockBuilder>> bbs;
-    DefMap<SpvId> params;
-};
-
 struct Builtins {
     SpvId workgroup_size;
     SpvId num_workgroups;
@@ -71,21 +51,7 @@ struct ImportedInstructions {
     explicit ImportedInstructions(FileBuilder&);
 };
 
-struct FileBuilder : public builder::SpvFileBuilder {
-    explicit FileBuilder(CodeGen* cg);
-    FileBuilder(const FileBuilder&) = delete;
-
-    CodeGen* cg;
-
-    std::unique_ptr<Builtins> builtins;
-    std::unique_ptr<ImportedInstructions> imported_instrs;
-
-    SpvId u32_t();
-    SpvId u32_constant(uint32_t);
-
-private:
-    SpvId u32_t_ { 0 };
-};
+struct BasicBlockBuilder;
 
 class CodeGen : public thorin::CodeGen, public thorin::Emitter<SpvId, ConvertedType, BasicBlockBuilder*, CodeGen> {
 public:
